CODAP-1352: move swapCategoriesNotification to data-display
V2 plugin-event compatibility (map): categorical-legend.tsx is shared
between graph and map tiles, but swapCategoriesNotification previously
lived in graph-notifications.ts and short-circuited for map tiles.
Move it to a new data-display-notifications.ts so the helper fires for
map's legend swap too — closing V2's analogous gap for map plugins.
(V2's legend Command at categories_view.js ~:120 had no
executeNotification block for either graph or map.)

- New: data-display/data-display-notifications.ts + .test.ts. No
  tile-type guard inside the helper — data-display must not import
  from graph or map; callers are responsible for invoking only where
  appropriate.
- Removed swapCategoriesNotification (and its tests) from
  graph-notifications.ts / .test.ts.
- Updated imports in categorical-legend.tsx and use-sub-axis.ts.
