+++ v3/src/components/axis/hooks/use-sub-axis.ts
@@ -9,7 +9,7 @@ import { isAliveSafe } from "../../../utilities/mst-utils"
 import { translate } from "../../../utilities/translation/translate"
 import { isVertical } from "../../axis-graph-shared"
 import { axisPlaceToAttrRole, kOther } from "../../data-display/data-display-types"
-import { swapCategoriesNotification } from "../../graph/graph-notifications"
+import { swapCategoriesNotification } from "../../data-display/data-display-notifications"
 import { useDataDisplayAnimation } from "../../data-display/hooks/use-data-display-animation"
 import { useDataDisplayModelContextMaybe } from "../../data-display/hooks/use-data-display-model"
 import { kDefaultFontHeight } from "../axis-constants"
