+++ v3/src/components/data-display/components/legend/categorical-legend.tsx
@@ -5,7 +5,7 @@ import { mstAutorun } from "../../../../utilities/mst-autorun"
 import { setSelectedCases, selectCases } from "../../../../models/data/data-set-utils"
 import { getTileModel } from "../../../../models/tiles/tile-model"
 import {axisGap} from "../../../axis/axis-types"
-import { swapCategoriesNotification } from "../../../graph/graph-notifications"
+import { swapCategoriesNotification } from "../../data-display-notifications"
 import { transitionDuration } from "../../data-display-types"
 import {useDataConfigurationContext} from "../../hooks/use-data-configuration-context"
 import { useDataDisplayModelContextMaybe } from "../../hooks/use-data-display-model"
