+++ v3/src/components/data-display/data-display-notifications.test.ts
@@ -1,0 +1,53 @@
+import { swapCategoriesNotification } from "./data-display-notifications"
+
+// V2 component-resource notifications carry the SC class name (`DG.GraphView` / `DG.MapView`)
+// in `values.type`; V3 adds the DI-convention name as `values.diType`. The mock resolves both
+// from the tile's content.type so the shared helpers can be exercised on either tile.
+jest.mock("../../models/tiles/tile-notifications", () => ({
+  updateTileNotification: jest.fn((updateType: string, values: any, tileModel: any) => {
+    if (!tileModel) return undefined
+    const tileType = tileModel.content?.type
+    const v2Type = tileType === "Graph" ? "DG.GraphView" : tileType === "Map" ? "DG.MapView" : ""
+    const diType = tileType === "Graph" ? "graph" : tileType === "Map" ? "map" : ""
+    return {
+      message: {
+        action: "notify",
+        resource: "component",
+        values: { operation: updateType, ...values, id: tileModel.id, type: v2Type, diType }
+      }
+    }
+  })
+}))
+
+describe("swapCategoriesNotification", () => {
+  it("emits 'swap categories' with the originating place (axis variant) for graph", () => {
+    const tile = { id: "GRAPH1", content: { type: "Graph" } } as any
+    const notification = swapCategoriesNotification(tile, "bottom")
+    expect(notification?.message.action).toBe("notify")
+    expect(notification?.message.resource).toBe("component")
+    expect(notification?.message.values.operation).toBe("swap categories")
+    expect(notification?.message.values.place).toBe("bottom")
+    expect(notification?.message.values.type).toBe("DG.GraphView")
+    expect(notification?.message.values.diType).toBe("graph")
+  })
+
+  it("emits 'swap categories' with place='legend' for the legend variant on graph", () => {
+    const tile = { id: "GRAPH1", content: { type: "Graph" } } as any
+    const notification = swapCategoriesNotification(tile, "legend")
+    expect(notification?.message.values.operation).toBe("swap categories")
+    expect(notification?.message.values.place).toBe("legend")
+  })
+
+  it("emits 'swap categories' for the legend variant on map (closes V2 map gap)", () => {
+    const tile = { id: "MAP1", content: { type: "Map" } } as any
+    const notification = swapCategoriesNotification(tile, "legend")
+    expect(notification?.message.values.operation).toBe("swap categories")
+    expect(notification?.message.values.place).toBe("legend")
+    expect(notification?.message.values.type).toBe("DG.MapView")
+    expect(notification?.message.values.diType).toBe("map")
+  })
+
+  it("returns undefined when the tile is missing", () => {
+    expect(swapCategoriesNotification(undefined, "left")).toBeUndefined()
+  })
+})
